Updated to Eco 9.6.
